Fix bugs

+++ src/main/java/com/gstu/services/RecordsTrainService.java
@@ -15,7 +15,9 @@ public class RecordsTrainService {
         }
 
 
-        trainDao.deleteById(id);
+        if (trainDao != null) {
+            trainDao.deleteById(id);
+        }
 
     }
 }
